half1/half1/requirements.txt
Flask==3.0.3
Flask-Cors==4.0.1
Pillow==10.4.0
requests==2.32.3
python-dotenv==1.0.1
openai==1.47.0
gTTS==2.5.3
ultralytics==8.0.196
opencv-python==4.8.1.78
numpy>=1.26.0
torch>=2.1.0
torchvision>=0.16.0
transformers>=4.35.0
accelerate>=0.24.0
pyttsx3==2.90
pytesseract==0.3.10
